\name{LFUrmutils-package}
\alias{LFUrmutils-package}
\alias{lrtest.fGARCH}
\docType{package}
\title{
LFU Risk Management Utilities
}
\description{
The package contains utility functions for the bachelor's seminar in risk management at the University of Innsbruck. In particular, it contains functions to compute and manipulate various univariate and multivariate volatility models. Future releases of this package may also provide functions to make outputs from other packages more comparable to those of this package.
}
\references{
Danielsson (2011). Financial Risk Forecasting. Wiley. Chichester.
Jorion (2007). Value at Risk, 3rd. McGraw-Hill. New York.
Ruppert and Matteson (2015). Statistics and Data Analysis for Financial Engineering, 2nd. Springer. New York.
}
\author{
Bernhard Eder
}
